+++ hbaselines/envs/mixed_autonomy/envs/av.py
@@ -132,7 +132,8 @@ class AVEnv(Env):
         return Box(
             low=-abs(self.env_params.additional_params['max_decel']),
             high=self.env_params.additional_params['max_accel'],
-            shape=(self.num_rl,))
+            shape=(self.num_rl,),
+            dtype=np.float32)
 
     @property
     def observation_space(self):
@@ -144,7 +145,8 @@ class AVEnv(Env):
         return Box(
             low=-float('inf'),
             high=float('inf'),
-            shape=(self.num_rl * (1 + 4 * max_lanes),))
+            shape=(self.num_rl * (1 + 4 * max_lanes),),
+            dtype=np.float32)
 
     def _apply_rl_actions(self, rl_actions):
         """See class definition."""
